refactor: split db.ts into domain files (barrel re-export)
Pure structural refactor: src/lib/db.ts (~2,300 lines, 129 exports) is
now a barrel that re-exports from domain files in src/lib/db/. No
behavior changes, no caller changes.

Domain files:
- profiles.ts            — current/profile fetch + update
- account.ts             — calendar token, version pings, account deletion
- events.ts              — events, saved events, people-down, crew pool, social signal
- event-comments.ts      — comments scoped to events
- friendships.ts         — friends, pending requests, friend links, suggestions, search
- checks.ts              — interest checks, co-authors, shared, hidden, archived, left
- check-comments.ts      — comments scoped to interest checks
- squads.ts              — squads, logistics, date confirms, join requests
- messages.ts            — squad chat messages + image uploads
- notifications.ts       — notifications + unread counts + read cursors
- polls.ts               — squad polls (text/dates/when/availability)
- blocks.ts              — block/unblock + report content
- api-base.ts            — shared API_BASE constant (avoids barrel import cycles)

Notes:
- All `import { foo } from "@/lib/db"` and `import * as db from "@/lib/db"`
  imports keep working unchanged.
- API_BASE is hoisted into a small api-base.ts so domain files don't have
  to import from the barrel (prevents circular imports). The barrel still
  re-exports it for callers.
- COMMENT_USER_COLS is duplicated in check-comments.ts and event-comments.ts
  rather than shared, since these two modules use the same check_comments
  table but logically belong to different domains.
- fetchEventUsersBatch (private helper) lives in events.ts and is called
  by both getPeopleDownBatch and getCrewPoolBatch — both crew-pool and
  saved-events functions live in events.ts to keep the helper local.

Verification:
- npx tsc --noEmit passes clean
- npm run test:e2e: same pass/fail set as origin/main (the two failing
  notification-routing tests fail on baseline too — pre-existing strict-
  mode selector collision unrelated to this refactor)
