+++ src/lib/db/api-base.ts
@@ -1,0 +1,2 @@
+/** Prefix for fetch('/api/...') calls. Empty in web, full URL in Capacitor. */
+export const API_BASE = process.env.NEXT_PUBLIC_API_BASE || '';
